2023.11.07 북마크 리사이클러뷰 2차 구현
리사이클러뷰 수정
