+++ app/src/main/java/com/example/memorycat/TodayWordStudyActivity.kt
@@ -1,5 +1,6 @@
 package com.example.memorycat
 
+import android.graphics.Color
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 
@@ -7,5 +8,9 @@ class TodayWordStudyActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_todayword_study)
+
+        //image_view.setColorFilter(Color.parseColor("#D2D2D20"))
+        //image_view.imageTintList.valueOf(Color.parseColor("#D2D2D2"))
+
     }
 }
