+++ app/src/main/res/layout/item_bookmark.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+xmlns:android="http://schemas.android.com/apk/res/android"
+android:id="@+id/item_root"
+android:layout_width="350dp"
+android:layout_height="wrap_content"
+android:orientation="horizontal"
+android:layout_gravity="center_horizontal"
+android:layout_marginBottom="15dp"
+android:padding="16dp"
+android:background="@drawable/rec_background">
+<LinearLayout
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginLeft="15dp"
+    android:orientation="vertical">
+    <TextView
+        android:id="@+id/nthWord"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:textStyle="bold"
+        android:textSize="30sp"
+        android:text="Cat"/>
+
+    <TextView
+        android:id="@+id/nthMean"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
+        android:layout_weight="1"
+        android:textSize="14sp"
+        android:textColor="@color/graydark"
+        android:text="1. 명사 고양이\n2. 명사 고양잇과 동물" />
+</LinearLayout>
+<LinearLayout
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginLeft="100dp"
+    android:orientation="horizontal">
+    <ImageView
+        android:id="@+id/bookmarkVoice"
+        android:layout_width="30dp"
+        android:layout_height="30dp"
+        android:src="@drawable/voice"/>
+    <ImageView
+        android:id="@+id/bookamrkBookmark"
+        android:layout_width="30dp"
+        android:layout_height="30dp"
+        android:src="@drawable/peow"/>
+</LinearLayout>
+</LinearLayout>
